Add DocumentPackagesProvider
