+++ Modules/Contracts/Generators/ProjectGeneratorContract/Sources/ProjectGeneratorContract/DocumentPackagesProviderContract.swift
@@ -1,0 +1,5 @@
+import PhoenixDocument
+
+public protocol DocumentPackagesProviderProtocol {
+    func packages(for document: PhoenixDocument) -> [PackageWithPath]
+}
