+++ Modules/Generators/ProjectGenerator/Sources/ProjectGenerator/Protocols/ComponentPackagesProviderProtocol.swift
@@ -1,15 +1,6 @@
 import PhoenixDocument
+import ProjectGeneratorContract
 import SwiftPackage
-
-public struct PackageWithPath: Equatable {
-    public let package: SwiftPackage
-    public let path: String
-
-    public init(package: SwiftPackage, path: String) {
-        self.package = package
-        self.path = path
-    }
-}
 
 public protocol ComponentPackagesProviderProtocol {
     func packages(for component: Component,
